[LH-159] Reduce number of adapter for auth #22
